### 2022-07-01 #4536 @editorialbot
[PRE REVIEW]: SSMSE: An R package for Management Strategy Evaluation with Stock Synthesis Operating Models
**Submitting author:** @k-doering-NOAA (<a href="http://orcid.org/0000-0002-0396-7044">Kathryn Doering</a>)
**Repository:** https://github.com/nmfs-fish-tools/ssmse
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.6
