### 2022-07-01 #4536 @Kevin-Mattheus-Moerman
@k-doering-NOAA thanks for this submission. I am a JOSS AEiC and here to help process initial steps.

- [ ] Can you check those potentially missing/invalid DOIs? :point_up: You can check the references gain by calling `@editorialbot check references`.

- [ ] There is a paper compilation error:

```
Problem with ORCID (00000-0002-0396-7044) for Kathryn Doering. ORCID looks to be the wrong length (Theoj::Error)
```
Looks like it should have 4 zeros at the start, not 5: https://orcid.org/0000-0002-0396-7044
Can you fix this and try to compile the paper again? You can call `@editorialbot generate pdf` here.
