### 2022-07-05 #4536 @danielskatz
👋 @k-doering-NOAA - note that your paper does not compile. As @Kevin-Mattheus-Moerman says, you at least need to remove the extra 0 in an ORCID. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
