+++ JOSS/paper.bib
@@ -32,7 +32,7 @@
 
   year = {2006},
 
-  doi = {10.1577/1548-8446(2006)31\[590:TCFIE\]2.0.CO;2},
+  doi = {10.1577/1548-8446(2006)31{\[}590:TCFIE{\]}2.0.CO;2},
 
   publisher={Wiley Online Library}
 
@@ -96,7 +96,9 @@ eprint = {https://www.science.org/doi/pdf/10.1126/science.1149016}
 
   year={2018},
 
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  
+  doi={10.1002/fsh.10131},
 
 }
 
@@ -143,7 +145,7 @@ year = {2020},
 
 issn = {0165-7836},
 
-url = {https://doi.org/10.1016/j.fishres.2020.105725},
+doi = {10.1016/j.fishres.2020.105725},
 
 author = {Jacob R. Krause and Joseph E. Hightower and Stephen J. Poland and Jeffrey A. Buckel},
 
@@ -181,7 +183,7 @@ year = {2021},
 
 issn = {0165-7836},
 
-url = {https://doi.org/10.1016/j.fishres.2020.105854},
+doi = {10.1016/j.fishres.2020.105854},
 
 author = {Marvin M. Mace and Kathryn L. Doering and Michael J. Wilberg and Amy Larimer and Frank Marenghi and Alexei Sharov and Mitchell Tarnowski}
 
@@ -337,7 +339,7 @@ year = {2022},
 
 issn = {0165-7836},
 
-url = {https://doi.org/10.1016/j.fishres.2022.106229},
+doi = {10.1016/j.fishres.2022.106229},
 
 author = {Éva E. Plagányi and Laura K. Blamey and Jacob G.D. Rogers and Vivitskaia J.D. Tulloch},
 
@@ -381,7 +383,7 @@ year = {2022},
 
 issn = {0165-7836},
 
-url = {https://doi.org/10.1016/j.fishres.2021.106180},
+doi = {10.1016/j.fishres.2021.106180},
 
 author = {Paul M. Regular and Alejandro D. Buren and Karen S. Dwyer and Noel G. Cadigan and Robert S. Gregory and Mariano Koen-Alonso and Rick M. Rideout and Gregory J. Robertson and Matthew D. Robertson and Garry B. Stenson and Laura J. Wheeland and Fan Zhang},
 
@@ -475,7 +477,7 @@ keywords = {Northern cod, Prey limitations, Body condition, Starvation, Mortalit
 
   year={2020},
   
-  url={https://doi.org/10.1111/faf.12480},
+  doi={10.1111/faf.12480},
 
   publisher={Wiley Online Library}
 
