### 2022-07-06 13c7386
fix: rm escape char
