+++ JOSS/paper.bib
@@ -32,7 +32,7 @@
 
   year = {2006},
 
-  doi = {10.1577/1548-8446(2006)31{\[}590:TCFIE{\]}2.0.CO;2},
+  doi = {10.1577/1548-8446(2006)31[590:TCFIE]2.0.CO;2},
 
   publisher={Wiley Online Library}
 
