### 2022-07-18 #4536 @k-doering-NOAA
@danielskatz and @Kevin-Mattheus-Moerman, just wondering if either of you had a suggestion of what to do about the one DOI I cannot figure out how to make valid? (see [previous post in thread](https://github.com/openjournals/joss-reviews/issues/4536#issuecomment-1176336065))
