+++ JOSS/paper.bib
@@ -32,7 +32,7 @@
 
   year = {2006},
 
-  doi = {10.1577/1548-8446{\%28}2006{\%29}31{\%5B}590{\%3A}TCFIE{\%5D}2.0.CO{\%3B}2},
+  doi = {10.1577/1548-8446{%28}2006{%29}31{%5B}590{%3A}TCFIE{%5D}2.0.CO{%3B}2},
 
   publisher={Wiley Online Library}
 
