### 2022-07-28 86f2e8b
test double curly braces in problem bib entry
