### 2022-07-28 #4536 @k-doering-NOAA
@lorenzo-rovigatti thanks for the suggestion! I gave it a try, (with a few variations of using escapes and grouping using curly braces) but the editorial bot is still marking the DOI as invalid
