### 2022-08-10 #4536 @k-doering-NOAA
I forgot to comment again about the DOI - my colleague thought what might be happening is that the "doi" was considered a doi via an old standard, but is no longer an acceptable format, which is why it is being marked as invalid now. However, the journal article does not offer an updated DOI, so I'm not sure where to go from here, maybe @xuanxu has some suggestions?
