### 2022-08-10 #4536 @xuanxu
DOI issue should be fixed now
