### 2022-08-10 #4536 @xuanxu
The `[`and `]` characters were causing the DOI url being invalid, now they are escaped before calling the DOI API :+1:
