### 2022-08-16 #4536 @Ecosantos
@sbenthall,  Thanks for the invitation but, I am not really sure if I can review this submission. Fishery stock assessment has some particular nuances out of my expertise.

I am happy to review anything related to population assessment of non-fishery stocks in the future.
