### 2022-08-18 #4536 @sbenthall
@hvoltbb Would you be able to review this submission to JOSS?
