### 2022-10-18 e95e6be
add workflow dispatch for calc coverage job
