### 2022-11-15 #4937 @seananderson
@sbenthall Regarding potential conflicts of interest: I have published in the last ~3 years with K. Marshal but on work unrelated to this. I've published with several of the co-authors at ~6 years ago on work related to this paper (ss3sim), but that should be outside the 4-year window. I also have some level of ongoing collaboration with co-authors K. Johnson, I. Taylor, and C. Wetzel but that is unrelated to this paper or topic and would not impact my impartial scientific judgment or evaluation.
