### 2022-12-02 #4937 @sbenthall
Hello. @Kevin-Mattheus-Moerman has weighed in on this.

@quang-huynh We can waive your COI in this case. Please proceed with your review!

@seananderson Unfortunately, we cannot waive your COI in this case. Thanks so much for reporting it. We will look for a different reviewer to replace you. If you could recommend anybody, that would be very helpful.
