### 2022-12-02 #4937 @seananderson
OK. Other recommendations that may be less likely to have conflicts of interest:
- [Laurence Kell](https://www.imperial.ac.uk/people/l.kell) or any co-authors of FLR: https://doi.org/10.1093/icesjms/fsm012 (should be fewer conflicts of interest given the European origin)
- [Cathy Dichmont](https://dichmont.weebly.com/) or [Natalie Dowling](https://people.csiro.au/D/N/Natalie-Dowling) from Australia
