### 2022-12-16 #4937 @sbenthall
Dear @ejardim,

I am reaching out as an Editor of the Journal of Open Source Software, an on-line journal of scientific research software.

The submission tracked in this issue is from a team led by Kathryn Doering (@k-doering-NOAA) and is an R package about fishery stock assessment. We have been struggling to find qualified reviewers as this is a specialized area. You were suggested as a reviewer by Sean Anderson, who was not able to review the submission because of a conflict of interest.

On behalf of the journal, I would like to ask you if you'd be willing to review this submission. The process takes place via GitHub and I'll be happy to walk you through it.

Best regards,

Sebastian Benthall @sbenthall
