### 2022-12-20 #4937 @quang-huynh
**Here is my written review:**

The authors present SSMSE, a R package that facilitates use of SS3, an executable software package, for management strategy evaluation, a simulation exercise which necessitates iterative application of SS3. I am familar with SS3 and a somewhat frequent user of the software.

### Manuscript comments

Line 20: Recommend "..that can dynamically..." It is also possible to compare static harvest strategies (fixed catches) in MSE.
Line 22: ..may specify how a stock assessment model is configured..
Line 24: I recommend "hypothesized" over "true". If we knew the true dynamics, we wouldn't need to generate multiple OMs and evaluate misspecification.

Figure 3. Expand 'User inputs' and 'Models' bubbles to include setup of management procedures and operating models with `create_om_list` and `sample_struct`.
Figure 4. Specify that this figure describes the case study in the caption. Also state the assumption of constant M in all EMs.

The paper lists five types of uncertainty that should be considered in MSE. The example demonstrates 1 and 2. Implicitly, one conditions a different SS3 model in advance to address model uncertainty (point 3). I couldn't figure out how to specify catch bias and implementation error in the operating model? (Bullets 4 and 5). I think I'd have to set up a model with these adjustments in the control or forecast file? I'm not sure. Example code in the manual would be useful to cover these last two bullets.

### Comments related to checklist

**Reproducibility**
I ran the script in: https://nmfs-fish-tools.github.io/SSMSE/manual/M-case-study-ex.html which I believe replicates the cod case study. No problems or convergence issues detected although I only ran 5 iterations instead of 100 due to runtime. The ggplot code to generate Figure 5 should be added on that web page.

**Functionality documentation**
Projected dynamics in the operating model can be tinkered with a list, with a template provided by `SSMSE::create_future_om_list`. However, there's no documentation in the help file on what information is needed and what options are possible. What are the required values for 'pars', 'scen', 'pattern', and 'input' entries in the list? I can figure it out with pattern recognition from the examples but that limits me to the case study.

Intimate familiarity with the SS3 model is a prerequisite to use SSMSE. To adjust natural mortality, I needed to know that the corresponding internal parameter name in this cod model is "NatM_p_1_Fem_GP_1". State where to find this in the r4ss output.

`create_sample_struct` seems to create a list that specifices the future data observations that will be simulated for the EM. State the options (catch, index, ages, etc) in the help file.

Examples can be documented in the help files with URL links to the online manual.

### Additional nice things to add

A simple wrapper function to generate the file structure needed for the analysis (model_runs, figures, input_models) would be helpful.

Due to long runtime, I recommend progress bars for `run_SSMSE` to inform users of elapsed and projected time. See the pbapply R package.

Are catch limits the only type of management advice that can be tested with SSMSE? Size limits are used to manage U.S. fisheries but I'm not sure if it can be implemented here?
