+++ .github/workflows/update-bookdown.yml
@@ -1,64 +1,16 @@
-# Workflow derived from https://github.com/r-lib/actions/tree/master/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
+# updates exiting bookdown for a repository
+# deploys to a branch gh-pages in the same repository
+name: call-update-bookdown
+# on specifies the build triggers. See more info at https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows
 on:
+# this workflow runs on pushes to main or master
   push:
     branches: main
-    paths: 'inst/bookdown/**'
-  workflow_dispatch:
-name: update-bookdown
-
+  workflow-dispatch:
 jobs:
-  bookdown:
-    runs-on: ubuntu-latest
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-    steps:
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-pandoc@v1
-
-      - uses: r-lib/actions/setup-r@v1
-        with:
-          use-public-rspm: true
-
-     # - uses: r-lib/actions/setup-renv@v1
-     
-      - name: Query dependencies
-        run: |
-          install.packages('devtools')
-          install.packages('remotes')
-          saveRDS(devtools::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
-          writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
-        shell: Rscript {0}
-
-      - name: Install system dependencies
-        if: runner.os == 'Linux'
-        run: |
-          while read -r cmd
-          do
-            eval sudo $cmd
-          done < <(Rscript -e 'writeLines(remotes::system_requirements("ubuntu", "20.04"))')
-      - name: Install dependencies
-        run: |
-          devtools::install_deps(dependencies = TRUE)
-          devtools::install_local(".")
-        shell: Rscript {0}
-      
-      - name: install extra pkgs (suggests or used in rmd files)
-        run: Rscript -e 'install.packages(c("bookdown"))'
-
-      - name: Cache bookdown results
-        uses: actions/cache@v2
-        with:
-          path: inst/bookdown/_bookdown_files
-          key: bookdown-${{ hashFiles('**/*Rmd') }}
-          restore-keys: bookdown-
-
-      - name: Build site
-        run: Rscript -e 'bookdown::render_book("inst/bookdown", output_dir = "../../docs/manual", quiet = TRUE)'
-
-      - name: Deploy to GitHub pages 🚀
-        uses: JamesIves/github-pages-deploy-action@4.1.4
-        with:
-          branch: gh-pages
-          folder: docs/
+  call-workflow:
+    uses: nmfs-fish-tools/ghactions4r/.github/workflows/update-bookdown.yml@refresh-bookdown
+      with:
+        bookdown-input: 'inst/bookdown'
+        bookdown-output_dir: '../../docs/manual'
+        deployment_dir: 'docs'
