+++ .github/workflows/update-bookdown.yml
@@ -5,12 +5,12 @@ name: call-update-bookdown
 on:
 # this workflow runs on pushes to main or master
   push:
-    branches: main
+    branches: test-bookdown
   workflow_dispatch:
 jobs:
   call-workflow:
     uses: nmfs-fish-tools/ghactions4r/.github/workflows/update-bookdown.yml@refresh-bookdown
     with:
-      bookdown-input: 'inst/bookdown'
-      bookdown-output_dir: '../../docs/manual'
+      bookdown_input: 'inst/bookdown'
+      bookdown_output_dir: '../../docs/manual'
       deployment_dir: 'docs'
