### 2023-06-15 65ab941
uses same workflow but from ss3sim
