### 2023-06-27 nmfs-fish-tools/SSMSE#148 @iagomosqueira
Running example in JOSS/paper.md
As part of the [review of the submission to JOSS](https://github.com/openjournals/joss-reviews/issues/4937), I have tried to run the example [linked here](https://github.com/nmfs-fish-tools/SSMSE/blob/main/JOSS/paper.md?plain=1#L285). But after quite a few hours on a large server, it is still unfinished.

Could maybe a simpler example be set up for demonstration purposes?
