### 2023-06-27 nmfs-fish-tools/SSMSE#148 @iagomosqueira
I would suggest this faster example is turned into an R script and linked in the JOSS paper so anybody looking into it can run it.
