### 2023-07-01 #4937 @iagomosqueira
# Review

The authors present SSMSE as package to conduct a subset of Management Strategy Evaluation (MSE) analysis in which the Operating Model (OM) is constructed around an stock assessment developed using the Stock Synthesis software. I am a MSE practitioner, familiar with the use of Stock Synthesis as a tool for conditioning OMs, and a developer of one of the MSE toolsets cited in the paper, [FLR](https://flr-project.org).

## Package

- Packages 'foreach' and 'doParallel' are only 'Suggests' in DESCRIPTION, but `foreach()` is used in package code, although its presence is checked. If option 'parallel=TRUE' is chosen, the code sets up a parallel cluster using `parallel::makeCluster`, when other backends could be employed, e.g. [doFuture](https://dofuture.futureverse.org/). An alternative option would be to import the 'foreach' package, and leave the setup to the user. For example, a Linux user would choose a 'FORK' cluster rather than the default 'PSOCK', for efficiency reasons.

- Can the functions source in the example from [here](https://raw.githubusercontent.com/k-doering-NOAA/ssmse-afs/master/code/get_metrics.R) be added to the package?

## Checklist

### General checks

- **Repository:** The package includes binary executables for which the [source code location](https://github.com/nmfs-stock-synthesis/stock-synthesis) is not specified. I could not find any way of checking what version of Stock Synthesis is included in the package.

- **Reproducibility:** The example run takes a very long time (days according to the author). I run it without problems but with only 5 iterations. I suggest a shorter example, even if less complete, so that an user can run it in a reasonable time frame, e.g. less than 2 hours. The author recommended the example contained in the README.md file, but it is a bit too limited.

### Documentation

- **Functionality documentation:** The set of available management procedures appears to be limited to those based on running SS as estimation method. Is the only way to run an alternative MP to write a function to takes SS dat and ctl files and input? A bfrief example would be an useful way of backing up the possible extensibility of the package.

### Software paper

- **State of the field:** I find the comparison with both OpenMSE and FLR too centered on their ability ot build MSEs around Stock Synthesis, ignoring that they offer a wider range of procedures and models.

- Line 115: "*but offer limited capacity to use existing stock assessment products created using SS3 for OM development*". SS3-based OMs are being used in the FLR platform in at least two MSE exercises: IOTC albacore and swordfish tuna (Sharma et al, 2020; [Brunel and Mosqueira, 2023] (https://iotc.org/sites/default/files/documents/2023/03/IOTC-2023-WPM14MSE-05_0.pdf)).

- **References:** Authors in entry *a4amsecite* should be corrected:

``` bibtex
author = {Jardim, JE and Scott, F and Mosqueira, I and Citores, L and Devine, J and Fischer, S and Ibaibarriaga, L and Mannini, A and Millar, C and Miller, D and Minto, C and De Oliveira, J and Osio, GC and Urtizberea, A and Vasilakopoulos, P and Kell, LT}
```
