### 2023-07-05 8b461cf
include info about SS3 binary
Co-authored-by: Nathan Vaughan <nathanvaughan-NOAA@users.noreply.github.com>
