### 2023-07-05 619e729
add parallel packages to imports instead of suggests
