+++ NAMESPACE
@@ -14,7 +14,6 @@ export(run_SSMSE)
 export(run_ss_model)
 export(set_MSE_seeds)
 import(ggplot2)
-import(parallel)
 import(r4ss)
 importFrom(assertive.base,assert_is_identical_to_true)
 importFrom(magrittr,"%>%")
