### 2023-07-05 7d25025
add some more info about other mse pkgs
