### 2023-07-13 97b2e34
add example perf. metrics Functions from https://raw.githubusercontent.com/k-doering-NOAA/ssmse-afs/master/code/get_metrics.R
