### 2023-07-13 98052da
partial: add docs, generalize perf. metrics
