+++ R/example_perf_metrics.R
@@ -1,29 +1,74 @@
-#' Calculate total catch over a range of years
+#' Example Performance Metric: Calculate total catch over a range of years
 #'
 #' @param datfile Path to the Stock Synthesis data file containing catch
 #' @param yrs A vector containing a range of years.
-get_total_catch<- function(datfile, yrs = 101:120) {
+#' @return The total catch, a number.
+#' @export
+#' @examples \dontrun{
+#'   total_catch <- get_total_catch(datfile = "ss3model/dat.ss", yrs = 25:100)
+#'   total_catch
+#' }
+get_total_catch<- function(datfile, yrs) {
    dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
    catch_sum <- sum(dat$catch[dat$catch$year %in% yrs, "catch"])
 }
 
-get_avg_catch<- function(datfile, yrs = 101:120) {
+#' Example Performance Metric: Calculate average catch over a range of years
+#' This is an example performance metric that will
+#'  calculate average catch over a range of years.
+#' Note that all fleets will be grouped together.
+#' @inheritParams get_total_catch
+#' @return The average catch, a number.
+#' @export
+#' @example \dontrun{
+#' avg_catch <- function(datfile = "ss3model/dat.ss", yrs = 30:75)
+#' avg_catch
+#' }
+get_avg_catch<- function(datfile, yrs) {
   dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
   catch_avg <- mean(dat$catch[dat$catch$year %in% yrs, "catch"])
 }
 
-get_catch_sd <- function(datfile, yrs = 101:120) {
+#' Example Performance Metric: Calculate Standard Deviation of Catch
+#' 
+#' Calculates the standard deviation of the catch across a range of years
+#' Does not take into account multiple fleets
+#' @inheritParams get_total_catch
+#' @return The catch standard deviation, a number.
+#' @export
+#' @example \dontrun{
+#'   catch_sd <- get_catch_sd(datfile = "mod/dat.ss", yrs = c(20:50, 75:100))
+#'   catch_sd
+#' }
+get_catch_sd <- function(datfile, yrs) {
   dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
   catch_var <- sd(dat$catch[dat$catch$year %in% yrs, "catch"])
 }
 
-get_catch_cv <- function(datfile, yrs = 101:120) {
+#' Example Performance Metric: Calculate the coefficient of variation of catch
+#' 
+#' Calculates the coefficient of variation of catch over a range of years. Note that
+#' this does not account for multiple fleets.
+#' @inheritParams get_total_catch
+#' @return The catch coefficient of variation, a number.
+#' @export
+get_catch_cv <- function(datfile, yrs) {
   dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
   catch <- dat$catch[dat$catch$year %in% yrs, "catch"]
   catch_var <- sd(catch)/mean(catch)
 }
 
-get_SSB_avg <- function(summary, min_yr = 175, max_yr = 200) {
+#' Example Performance Metric: calculate the average SSB over a range of years
+#' 
+#' Calculates the average SSB for each iteration of each scenario in 
+#' a run of SSMSE
+#'  
+#' @param summary Summary returned from running `run_SSMSE()`
+#' @param min_year The first year to include in the average
+#' @param max_year The last year to include in the average
+#' @export
+#' @return A tibble containing the Average SSB by iteration and scenario.
[…]
