### 2023-07-14 bb770f3
example performance metrics, add tests, update docs.
