+++ R/example_perf_metrics.R
@@ -1,42 +1,51 @@
 #' Example Performance Metric: Calculate total catch over a range of years
 #'
+#' Example performance metric that calculates total catch over a range of years 
+#'  in a Stock Synthesis data file. This function aggregates catch across
+#'  fleets, so may not be appropriate for models with multiple fleets.
 #' @param datfile Path to the Stock Synthesis data file containing catch
-#' @param yrs A vector containing a range of years.
-#' @return The total catch, a number.
+#' @param yrs A vector containing a range of years. Years are as defined in the
+#'  Stock Synthesis data file.
+#' @returns The total catch, a number.
 #' @export
 #' @examples \dontrun{
 #'   total_catch <- get_total_catch(datfile = "ss3model/dat.ss", yrs = 25:100)
 #'   total_catch
 #' }
-get_total_catch<- function(datfile, yrs) {
+get_total_catch <- function(datfile, yrs) {
    dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
    catch_sum <- sum(dat$catch[dat$catch$year %in% yrs, "catch"])
 }
 
 #' Example Performance Metric: Calculate average catch over a range of years
-#' This is an example performance metric that will
-#'  calculate average catch over a range of years.
-#' Note that all fleets will be grouped together.
+#'
+#' Example performance metric that calculates average catch over a range of
+#'  years in a Stock Synthesis data file. This function aggregates across
+#'  fleets, so may not be appropriate for models with multiple fleets.
 #' @inheritParams get_total_catch
-#' @return The average catch, a number.
+#' @returns The average catch, a number.
 #' @export
-#' @example \dontrun{
+#' @examples
+#' \dontrun{
 #' avg_catch <- function(datfile = "ss3model/dat.ss", yrs = 30:75)
 #' avg_catch
 #' }
-get_avg_catch<- function(datfile, yrs) {
+get_avg_catch <- function(datfile, yrs) {
   dat <- r4ss::SS_readdat(datfile, verbose = FALSE)
   catch_avg <- mean(dat$catch[dat$catch$year %in% yrs, "catch"])
 }
 
 #' Example Performance Metric: Calculate Standard Deviation of Catch
-#' 
-#' Calculates the standard deviation of the catch across a range of years
-#' Does not take into account multiple fleets
+#'
+#' Example performance metric that calculates the standard deviation of catch 
+#'  over a range of years in a Stock Synthesis data file. This function 
+#'  aggregates across fleets, so may not be appropriate for models with multiple
+#'  fleets.
 #' @inheritParams get_total_catch
-#' @return The catch standard deviation, a number.
+#' @returns The catch standard deviation, a number.
 #' @export
-#' @example \dontrun{
+#' @examples 
+#' \dontrun{
 #'   catch_sd <- get_catch_sd(datfile = "mod/dat.ss", yrs = c(20:50, 75:100))
 #'   catch_sd
 #' }
@@ -46,12 +55,19 @@ get_catch_sd <- function(datfile, yrs) {
 }
 
 #' Example Performance Metric: Calculate the coefficient of variation of catch
-#' 
-#' Calculates the coefficient of variation of catch over a range of years. Note that
-#' this does not account for multiple fleets.
+#'
+#' Example performance metric that calculates the coefficient of variation (CV)
+#'  of catch over a range of years in a Stock Synthesis data file. This function 
[…]
