+++ man/get_catch_sd.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/example_perf_metrics.R
+\name{get_catch_sd}
+\alias{get_catch_sd}
+\title{Example Performance Metric: Calculate Standard Deviation of Catch}
+\usage{
+get_catch_sd(datfile, yrs)
+}
+\arguments{
+\item{datfile}{Path to the Stock Synthesis data file containing catch}
+
+\item{yrs}{A vector containing a range of years. Years are as defined in the
+Stock Synthesis data file.}
+}
+\value{
+The catch standard deviation, a number.
+}
+\description{
+Example performance metric that calculates the standard deviation of catch 
+ over a range of years in a Stock Synthesis data file. This function 
+ aggregates across fleets, so may not be appropriate for models with multiple
+ fleets.
+}
+\examples{
+\dontrun{
+  catch_sd <- get_catch_sd(datfile = "mod/dat.ss", yrs = c(20:50, 75:100))
+  catch_sd
+}
+}
