+++ man/get_total_catch.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/example_perf_metrics.R
+\name{get_total_catch}
+\alias{get_total_catch}
+\title{Example Performance Metric: Calculate total catch over a range of years}
+\usage{
+get_total_catch(datfile, yrs)
+}
+\arguments{
+\item{datfile}{Path to the Stock Synthesis data file containing catch}
+
+\item{yrs}{A vector containing a range of years. Years are as defined in the
+Stock Synthesis data file.}
+}
+\value{
+The total catch, a number.
+}
+\description{
+Example performance metric that calculates total catch over a range of years 
+ in a Stock Synthesis data file. This function aggregates catch across
+ fleets, so may not be appropriate for models with multiple fleets.
+}
+\examples{
+\dontrun{
+  total_catch <- get_total_catch(datfile = "ss3model/dat.ss", yrs = 25:100)
+  total_catch
+}
+}
