### 2023-07-14 b9a4a38
add detail to intro on general MSE tools
