### 2023-07-14 0c07321
reproducibility: add comment on sim time let potential readers of the paper know how they could reproduce the example.
