### 2023-07-21 #4937 @k-doering-NOAA
Thanks for the reviews, @iagomosqueira and @quang-huynh!

@sbenthall, yes, I am working with my co-authors on making changes and putting together responses. I think we should be able to finish within the next month, if that sounds reasonable.
