### 2023-07-21 a99510c
add code for plotting joss demo
