+++ inst/bookdown/02-ex-simple.Rmd
@@ -552,3 +552,198 @@ summary <- SSMSE::SSMSE_summary_all(dir = "model_runs", run_parallel = TRUE)
 # #check for errored iterations
 lapply(out, function(x) x$errored_iterations)
 ```
+
+To examine simulations for non-convergance, calculate, and plot performance metrics:
+
+```{r M-case-study-plot, eval=FALSE}
+# Look at results from all runs (100 each iter) ----
+
+# Load packages set options ----
+
+library(SSMSE)
+library(r4ss)
+library(dplyr)
+library(ggplot2)
+library(patchwork)
+# functions for convergence and performance metrics, get from other gh repo
+source("https://raw.githubusercontent.com/k-doering-NOAA/ssmse-afs/master/code/get_metrics.R")
+# path names ----
+mods_path <- "input_models"
+
+# define the scenarios ----
+scen_red_tide <- c("no-red-tide", "low-red-tide", "hi-red-tide")
+scen_HCR <- c("F-spr-30", "F-spr-45")
+
+scenarios <- data.frame(
+  scen_name = c(paste0(scen_red_tide, "-",scen_HCR[1]),
+                paste0(scen_red_tide, "-", scen_HCR[2])),
+  EM_path = rep(c(file.path(mods_path, scen_HCR)), times = c(3,3))
+)
+
+# get the files that had issuse running ----
+
+error_mods <- lapply(out, function(x) {
+  tmp <- x$errored_iterations
+  if(isTRUE(tmp == "No errored iterations")) {
+    tmp <- NULL
+  }
+  tmp
+}
+)
+
+error_mods_df <- do.call(bind_rows, error_mods)
+error_mods_key <- error_mods_df[,c("iteration", "scenario")]
+
+# remove the models with issues
+
+summary$ts <- dplyr::anti_join(summary$ts, error_mods_key)
+summary$scalar <- dplyr::anti_join(summary$scalar, error_mods_key)
+
+## check convergence ----
+
+check_scalar <- summary$scalar[,c("max_grad", "iteration", "scenario")]
+too_high_max_grad_key <- na.omit(summary$scalar[summary$scalar$max_grad>2, c("iteration", "scenario")])
+summary$ts <- dplyr::anti_join(summary$ts, too_high_max_grad_key)
+summary$scalar <- dplyr::anti_join(summary$scalar, too_high_max_grad_key)
+
+
+SSB_df <- check_convergence(summary, n_EMs = 6, max_yr = 150)
+summary(SSB_df$SSB_ratio)
+
+
+SSB_df# no params on bounds, there are some relatively low or high SSB's.
+
+# how many iterations per scenario are left? 
+
+n_iters_per_scen <- summary$scalar[summary$scalar$model_run == "cod_OM", c("iteration", "scenario")] %>% 
+  group_by(scenario) %>% 
+  summarize(n = n())
+write.csv(n_iters_per_scen, "model_runs/n_iter_per_scen.csv")
+
+# write problem scenarios to afile
+write.csv(too_high_max_grad_key, "model_runs/too_high_max_grad.csv")
+write.csv(error_mods_key, "model_runs/error_mods_key.csv")
+
+all_errors <- rbind(too_high_max_grad_key, error_mods_key)
+
+# calculate performance metrics ----
+# look at catch in OM from yrs 125:150
+OM_metrics <- NULL
+for (i in scenarios$scen_name) { # scenarios$scen_name to make general
+  
+  iterations <- list.dirs(file.path("model_runs", i), recursive = FALSE, full.names = FALSE)
+  # remove iterations that had errors/convergence issues
+  test_df <- data.frame( iteration = as.double(iterations), scenario = i)
+  test_df <- dplyr::anti_join(test_df, all_errors)
[…]
