### 2023-07-21 1452b0a
add links to manual to R documentation
