### 2023-07-21 b9448b5
add link to user manual from run_SSMSE roxygen.
