+++ R/runSSMSE.R
@@ -1,7 +1,8 @@
 #' run an MSE using SS OMs
 #'
 #' High level function to run a management strategy evaluation using Stock
-#' Synthesis as the Operating model(s)
+#' Synthesis as the Operating model(s). For more examples and information on how
+#' to use SSMSE, see the [SSMSE user manual](https://nmfs-fish-tools.github.io/SSMSE/manual/).
 #' @template run_SSMSE_scen_list
 #' @template seed_input
 #' @template future_om_list
