### 2023-07-21 46b8865
simplify link to user manual
