### 2023-07-25 nmfs-fish-tools/SSMSE#151 @k-doering-NOAA
Progress bar 
Suggested by Quang in JOSS review:

Due to long runtime, I recommend progress bars for run_SSMSE to inform users of elapsed and projected time. See the pbapply R package.
