### 2023-07-25 nmfs-fish-tools/SSMSE#152 @k-doering-NOAA
add more parallel options
Suggestion On Iago's SSMSE JOSS review:

Packages 'foreach' and 'doParallel' are only 'Suggests' in DESCRIPTION, but `foreach()` is used in package code, although its presence is checked. If option 'parallel=TRUE' is chosen, the code sets up a parallel cluster using `parallel::makeCluster`, when other backends could be employed, e.g. [doFuture](https://dofuture.futureverse.org/). An alternative option would be to import the 'foreach' package, and leave the setup to the user. For example, a Linux user would choose a 'FORK' cluster rather than the default 'PSOCK', for efficiency reasons.

@nathanvaughan-NOAA is working on some other parallel code for another project, so we can build upon his work after it is done.
