### 2023-07-26 583d934
Add startup message that includes the SS3 version
