### 2023-07-26 29498e4
Add link in the roxygen to ssmse manual, custom MPs
