### 2023-07-27 cde3ffc
add note about sourcing fxns
