### 2023-07-27 nmfs-fish-tools/SSMSE#151 @k-doering-NOAA
https://peter.solymos.org/pbapply/

This can probably be done when refactoring the parallel options (#152 ). In general, pbapply seems easy to add, but there is a note that it is a "little more complicated" for windows users. Also there is a note about the future backend: https://peter.solymos.org/pbapply/#working-with-a-future-backend
