### 2023-07-28 #4937 @k-doering-NOAA
@quang-huynh, thank you for the thorough review and helpful suggestions that improved our manuscript and documentation. Edits suggested from the review were made on the [`joss-review` branch in the SSMSE repository](https://github.com/nmfs-fish-tools/ssmse/tree/joss-review).  These will be merged into main after wrapping up the review process (unless there is a better time to merge these in). To see all differences from the main branch, the compare page is helpful: https://github.com/nmfs-fish-tools/SSMSE/compare/main...joss-review . Note that changes to the manual and readme will not be shown in the rendered readme and manual until the changes are pushed to main.

We hope that we have addressed your suggestions sufficiently to satisfy the checklist, but please let us know if there are additional changes we need to make.

> **Here is my written review:**
>
> The authors present SSMSE, a R package that facilitates use of SS3, an executable software package, for management strategy evaluation, a simulation exercise which necessitates iterative application of SS3. I am familar with SS3 and a somewhat frequent user of the software.
>
> ### Manuscript comments
> Line 20: Recommend "..that can dynamically..." It is also possible to compare static harvest strategies (fixed catches) in MSE. Line 22: ..may specify how a stock assessment model is configured.. Line 24: I recommend "hypothesized" over "true". If we knew the true dynamics, we wouldn't need to generate multiple OMs and evaluate misspecification.

Changes made in commit https://github.com/nmfs-fish-tools/SSMSE/commit/fb6b26f. Thank you for these helpful recommendations!

> Figure 3. Expand 'User inputs' and 'Models' bubbles to include setup of management procedures and operating models with `create_om_list` and `sample_struct`. Figure 4. Specify that this figure describes the case study in the caption. Also state the assumption of constant M in all EMs.

While we see value in adding `create_om_list` and `sample_struct` in Figure 3, we are concerned that adding them in will over complicate the figure. Instead, we opted for an additional sentence in the caption highlighting the two functions (change made in https://github.com/nmfs-fish-tools/SSMSE/commit/a00f1b6). We hope this addresses the suggestion still!

For Figure 4, we edited the caption to include that the figure describes the case study and that the EMs assume constant M (https://github.com/nmfs-fish-tools/SSMSE/commit/fb6b26f).

> The paper lists five types of uncertainty that should be considered in MSE. The example demonstrates 1 and 2. Implicitly, one conditions a different SS3 model in advance to address model uncertainty (point 3). I couldn't figure out how to specify catch bias and implementation error in the operating model? (Bullets 4 and 5). I think I'd have to set up a model with these adjustments in the control or forecast file? I'm not sure. Example code in the manual would be useful to cover these last two bullets.

Implementation error can be specified by using the `future_om_list`, as outlined in [the structure of future_om_list section](https://nmfs-fish-tools.github.io/SSMSE/manual/SSMSE.html#future-changes-to-the-operating-model-future) of the user manual. So, no need to add adjustments in the control or forecast file.

For catch bias, you could create a custom management procedure and add code there to create catch bias. There is an example of a custom management procedure in [the custom management procedure section of the SSMSE User Manual](https://nmfs-fish-tools.github.io/SSMSE/manual/advanced-options-use-a-custom-management-strategyprocedure.html).

> ### Comments related to checklist
> **Reproducibility** I ran the script in: https://nmfs-fish-tools.github.io/SSMSE/manual/M-case-study-ex.html which I believe replicates the cod case study. No problems or convergence issues detected although I only ran 5 iterations instead of 100 due to runtime. The ggplot code to generate Figure 5 should be added on that web page.
>

Good point, we added the code to create Figure 5 in https://github.com/nmfs-fish-tools/SSMSE/commit/a99510c

> **Functionality documentation** Projected dynamics in the operating model can be tinkered with a list, with a template provided by `SSMSE::create_future_om_list`. However, there's no documentation in the help file on what information is needed and what options are possible. What are the required values for 'pars', 'scen', 'pattern', and 'input' entries in the list? I can figure it out with pattern recognition from the examples but that limits me to the case study.
>
It’s a really good point - we added a link to the SSMSE user manual in the R help documentation (https://github.com/nmfs-fish-tools/SSMSE/commit/1452b0a).

> Intimate familiarity with the SS3 model is a prerequisite to use SSMSE. To adjust natural mortality, I needed to know that the corresponding internal parameter name in this cod model is "NatM_p_1_Fem_GP_1". State where to find this in the r4ss output.

We address the parameter names in the [structure of future_om_list section](https://nmfs-fish-tools.github.io/SSMSE/manual/SSMSE.html#future-changes-to-the-operating-model-future) of the SSMSE user manual, which reads:

 “The first list item is named “pars”. It contains a vector of parameter name(s) to apply the change to. The names should be the same as the names in `r4ss::SS_read_pars()`”

> `create_sample_struct` seems to create a list that specifices the future data observations that will be simulated for the EM. State the options (catch, index, ages, etc) in the help file.

> Examples can be documented in the help files with URL links to the online manual.
The options and a link to the SSMSE manual are now stated in the help file for `create_sample_struct` (https://github.com/nmfs-fish-tools/SSMSE/commit/1452b0a)

> ### Additional nice things to add
> A simple wrapper function to generate the file structure needed for the analysis (model_runs, figures, input_models) would be helpful.

We can see the appeal, but we are a little concerned about requiring people to use a particular way of organizing their code (`figures` and `input_models` in separate folders from where the `model_runs` take place). I’m not sure this workflow is the way everyone likes to work.

> Due to long runtime, I recommend progress bars for `run_SSMSE` to inform users of elapsed and projected time. See the pbapply R package.

Thanks for suggesting the {pbapply} package! The other reviewer suggested expanding the options for running in parallel, so we think adding progress bars would best be done when refactoring the code to add in more parallel options. We opened an issue for the progress bar: https://github.com/nmfs-fish-tools/SSMSE/issues/151

> Are catch limits the only type of management advice that can be tested with SSMSE? Size limits are used to manage U.S. fisheries but I'm not sure if it can be implemented here?

For now, yes, only catch limits can be used. Nathan has been working on adding size limits to SSMSE, but it is still in development. It can be implemented, but would require some careful changes and testing to the current codebase. We agree that this would be useful to add.
