### 2023-07-28 #4937 @k-doering-NOAA
@iagomosqueira, thank you so much for your review that improved our paper and source code. We hope that the edits we make are sufficient to fulfill the checklist, but please let us know if there are additional changes to make.

 Edits suggested from the review were made on the [`joss-review` branch in the SSMSE repository](https://github.com/nmfs-fish-tools/ssmse/tree/joss-review).  These will be merged into main after wrapping up the review process (unless there is a better time to merge these in). To see all differences from the main branch, the compare page is helpful: https://github.com/nmfs-fish-tools/SSMSE/compare/main...joss-review. Note that changes to the manual and readme will not be shown in the rendered readme and manual until the changes are pushed to main.

> # Review
> The authors present SSMSE as package to conduct a subset of Management Strategy Evaluation (MSE) analysis in which the Operating Model (OM) is constructed around an stock assessment developed using the Stock Synthesis software. I am a MSE practitioner, familiar with the use of Stock Synthesis as a tool for conditioning OMs, and a developer of one of the MSE toolsets cited in the paper, [FLR](https://flr-project.org).
>
> ## Package
> * Packages 'foreach' and 'doParallel' are only 'Suggests' in DESCRIPTION, but `foreach()` is used in package code, although its presence is checked. If option 'parallel=TRUE' is chosen, the code sets up a parallel cluster using `parallel::makeCluster`, when other backends could be employed, e.g. [doFuture](https://dofuture.futureverse.org/). An alternative option would be to import the 'foreach' package, and leave the setup to the user. For example, a Linux user would choose a 'FORK' cluster rather than the default 'PSOCK', for efficiency reasons.
Great points about the parallel code - we changed foreach, parallel, and doParallel to imports instead of suggests, and removed the checks for presence since they were no longer needed (done in commit https://github.com/nmfs-fish-tools/SSMSE/commit/619e729).

We agree that having more flexible parallel options would be helpful for the users of SSMSE. However, we think that making this change in the future would make the most sense. Nathan is working on parallel options for a different project, and we can apply his work to SSMSE at a later time, if this seems acceptable. We opened an issue (https://github.com/nmfs-fish-tools/SSMSE/issues/152).

In the meantime, there are additional ways to run parallel simulations using options outside of SSMSE; for instance, a user could use the furrr package to run scenarios in parallel (example: https://github.com/k-doering-NOAA/ssmse-afs/blob/master/code/run_parallel_example.R).

> * Can the functions source in the example from [here](https://raw.githubusercontent.com/k-doering-NOAA/ssmse-afs/master/code/get_metrics.R) be added to the package?

The functions were added in commits https://github.com/nmfs-fish-tools/SSMSE/commit/97b2e34, https://github.com/nmfs-fish-tools/SSMSE/commit/98052da, and https://github.com/nmfs-fish-tools/SSMSE/commit/bb770f3. Documentation and tests were added as well.  The example code was changed to make it clear the functions have been added to SSMSE (https://github.com/nmfs-fish-tools/SSMSE/commit/cde3ffc).

> ## Checklist
> ### General checks
> * **Repository:** The package includes binary executables for which the [source code location](https://github.com/nmfs-stock-synthesis/stock-synthesis) is not specified. I could not find any way of checking what version of Stock Synthesis is included in the package.

Thanks for pointing this out. We added some information to the readme on the version of the Stock Synthesis binaries (3.30.18) in commit https://github.com/nmfs-fish-tools/SSMSE/commit/8b461cf.
We also included a simple startup message to the user upon loading the package what the version of SS3 is (commit https://github.com/nmfs-fish-tools/SSMSE/commit/583d934).

> * **Reproducibility:** The example run takes a very long time (days according to the author). I run it without problems but with only 5 iterations. I suggest a shorter example, even if less complete, so that an user can run it in a reasonable time frame, e.g. less than 2 hours. The author recommended the example contained in the README.md file, but it is a bit too limited.
>
The goal of the example in the paper is to provide a use case for SSMSE rather than a simple example for a user to follow along with. Because the SSMSE examples run full assessment models, the run times are naturally long for a complete simulation study. We added a note to readers so they are aware of the long run times and that they can reduce the number of iterations to decrease runtime (https://github.com/nmfs-fish-tools/SSMSE/commit/0c07321).

We could potentially create a simpler example that avoids running Estimation Models for the assessment, but it would not demonstrate the full capabilities of SSMSE as well.

> ### Documentation
> * **Functionality documentation:** The set of available management procedures appears to be limited to those based on running SS as estimation method. Is the only way to run an alternative MP to write a function to takes SS dat and ctl files and input? A brief example would be an useful way of backing up the possible extensibility of the package.

There is an option to write your own management strategy that need not run SS3 as the estimation model, see the [custom management strategy section of the SSMSE user manual](https://nmfs-fish-tools.github.io/SSMSE/manual/advanced-options-use-a-custom-management-strategyprocedure.html). We added a link from the Roxygen documentation to the manual to make this option more discoverable to users (commit https://github.com/nmfs-fish-tools/SSMSE/commit/29498e4).
>
> ### Software paper
> * **State of the field:** I find the comparison with both OpenMSE and FLR too centered on their ability to build MSEs around Stock Synthesis, ignoring that they offer a wider range of procedures and models.
> * Line 115: "_but offer limited capacity to use existing stock assessment products created using SS3 for OM development_". SS3-based OMs are being used in the FLR platform in at least two MSE exercises: IOTC albacore and swordfish tuna (Sharma et al, 2020; [Brunel and Mosqueira, 2023]
(https://iotc.org/sites/default/files/documents/2023/03/IOTC-2023-WPM14MSE-05_0.pdf)).

Thanks for providing input. We respect the hard work that has gone into FLR and OpenMSE, and fully agree that they have extensive capabilities. The goal of SSMSE was Stock Synthesis-centric, but we want to highlight the capabilities that OpenMSE and FLR have, which are more general than SSMSE’s features. We made some edits to try to address this, but please let us know if it still doesn’t sound correct and feel free to suggest additional edits (edit made in commit https://github.com/nmfs-fish-tools/SSMSE/commit/7d25025).

We already have cited Sharma et al. 2020 in the paper, but for the references provided there was no mention of FLR - for example, we didn’t know that the Sharma et al. 2020 paper operating models had used FLR as a framework because there was no mention of FLR, unfortunately. We thought from reading the paper that analysts came up with their own bespoke frameworks for the cases using SS3 operating models.

We also think (but may have misunderstood) that SS3 is not used directly as an operating model in the FLR framework, but rather an SS3 run is converted to an FLR Operating model, which may result in some loss of complexity. While there are many situations where this is advantageous or sufficient, SSMSE took a different approach to fill a gap that we thought existed in generalized tools - we directly use SS3 as an operating model. We tried to explain this difference in the paper, but may have not worded it correctly (or perhaps we have misunderstood how FLR’s MSE package works). Please feel free to correct us if the language still does not accurately reflect how FLR’s MSE package works.

> * **References:** Authors in entry _a4amsecite_ should be corrected:
>
> ```
> author = {Jardim, JE and Scott, F and Mosqueira, I and Citores, L and Devine, J and Fischer, S and Ibaibarriaga, L and Mannini, A and Millar, C and Miller, D and Minto, C and De Oliveira, J and Osio, GC and Urtizberea, A and Vasilakopoulos, P and Kell, LT}
> ```

Thank you for catching this and apologies for the mistake - Change made in commit https://github.com/nmfs-fish-tools/SSMSE/commit/c57bc0c.
