### 2023-08-09 47445d9
style and docs: run devtools::document() and styler::style_pkg()
