### 2023-08-09 fd1414b
style and docs: run devtools::document() and styler::style_pkg()
