+++ README.Rmd
@@ -97,8 +97,8 @@ First, we will load the `SSMSE` package and create a folder in which to run the 
 devtools::install_local(".", upgrade = "never", quiet = TRUE, force = TRUE)
 ```
 ```{r Load-Packages, eval=TRUE}
-library(SSMSE) #load the package
-library(r4ss) #install using remotes::install_github("r4ss/r4ss)
+library(SSMSE) # load the package
+library(r4ss) # install using remotes::install_github("r4ss/r4ss)
 # library(foreach) #if using run_parallel = TRUE
 # library(doParallel) #if using run_parallel = TRUE
 ```
@@ -120,8 +120,10 @@ First, we identify where the base cod model is stored, modify it such that the s
 cod_mod_path <- system.file("extdata", "models", "cod", package = "SSMSE")
 # develop_OMs will save a model called "cod_SR_BH_steep_1" in the out_dir
 # specified
-develop_OMs(OM_name = "cod", out_dir = run_SSMSE_dir, par_name = "SR_BH_steep",
-            par_vals = 1, refit_OMs = FALSE, hess = FALSE)
+develop_OMs(
+  OM_name = "cod", out_dir = run_SSMSE_dir, par_name = "SR_BH_steep",
+  par_vals = 1, refit_OMs = FALSE, hess = FALSE
+)
 # OM model for scenario 2
 cod_1_path <- file.path(run_SSMSE_dir, "cod_SR_BH_steep_1")
 ```
@@ -134,7 +136,7 @@ First, we'll set up the list to add recruitment deviations in the projection per
 
 ```{r Setup-Recdevs}
 # Start from a list created by a helper function
-template_mod_change <- create_future_om_list() 
+template_mod_change <- create_future_om_list()
 # add recruitment deviations
 rec_dev_specify <- template_mod_change[[1]]
 rec_dev_specify$pars <- "rec_devs" # apply change to rec devs
@@ -173,8 +175,10 @@ mod_change_sel
 
 Finally, add these two changes together into an object to pass to `run_SSMSE`
 ```{r Fut-OM}
-future_om_list_recdevs_sel <- list(rec_dev_specify, 
-                                   mod_change_sel) 
+future_om_list_recdevs_sel <- list(
+  rec_dev_specify,
+  mod_change_sel
+)
 ```
 
 ### Add observation error through sampling from OM
@@ -216,8 +220,9 @@ Let's take a look at step 2 in the management procedure, which is implemented us
 ```{r See-Fore-Vals}
 fore <- r4ss::SS_readforecast(
   system.file("extdata", "models", "cod", "forecast.ss", package = "SSMSE"),
-  verbose = FALSE)
-fore$Forecast 
+  verbose = FALSE
+)
+fore$Forecast
 fore$Btarget
 ```
 `fore$Forecast = 3` means our forecasts from the assessment will use fishing mortality (F) to attmpt to achieve a relative (to unfished) spawning stock biomass. Based on `fore$Btarget`, the relative biomass target is 40% of unfished spawning stock biomass. Note also that the control rule `fore$BforconstantF` and `fore$BfornoF` values are set low to make it unlikely that they will be used (these parameters are used for a ramp harvest control rule, which we do not want to use here):
@@ -253,21 +258,22 @@ Now, we create a directory to store our results, and use `run_SSMSE` to run the 
 run_res_path <- file.path(run_SSMSE_dir, "results")
 dir.create(run_res_path)
 res <- run_SSMSE(
-    scen_name_vec = c("h-ctl", "h-1"),# name of the scenario
-    out_dir_scen_vec = run_res_path, # directory in which to run the scenario
-    iter_vec = c(5,5), # run with 5 iterations each
-    OM_name_vec = NULL, # specify directories instead
[…]
