### 2023-09-06 0bd6fb7
style and docs: run devtools::document() and styler::style_pkg()
