+++ tests/testthat/test-runSSMSE.R
@@ -134,38 +134,45 @@ test_that("run_SSMSE runs multiple iterations/scenarios and works with summary f
   summary <- SSMSE_summary_all(dir = new_temp_path, run_parallel = FALSE)
   expect_true(length(summary) == 3)
   # convergance check function
-  
+
   # calculate performance metrics
   # to use in the tests
-    tmp_dat <- r4ss::SS_readdat(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"))
-  #check dimensions
+  tmp_dat <- r4ss::SS_readdat(
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new")
+  )
+  # check dimensions
   # catch related
   tot_catch <- get_total_catch(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 101:106)
-  
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 101:106
+  )
+
   expect_length(tot_catch, 1)
   tot_catch_zero <- get_total_catch(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 1:25)
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 1:25
+  )
   expect_equivalent(tot_catch_zero, 0)
   avg_catch <- get_avg_catch(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 101:106)
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 101:106
+  )
   expect_length(avg_catch, 1)
   avg_catch_2_yrs <- get_avg_catch(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 26:27)
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 26:27
+  )
   compare_catch <- mean(tmp_dat$catch[tmp_dat$catch$year %in% c(26, 27), "catch"])
   expect_equivalent(avg_catch_2_yrs, compare_catch)
   var_sd_catch <- get_catch_sd(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 101:106)
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 101:106
+  )
   expect_length(var_sd_catch, 1)
   var_cv_catch <- get_catch_cv(
-    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"), 
-    yrs = 101:106)
+    file.path(new_temp_path, "H-scen-2", "1", "cod_OM", "data.ss_new"),
+    yrs = 101:106
+  )
   expect_length(var_cv_catch, 1)
   # SSB related
   ssb_avg <- get_SSB_avg(summary, min_yr = 101, max_yr = 106)
@@ -179,7 +186,7 @@ test_that("run_SSMSE runs multiple iterations/scenarios and works with summary f
     ssb_check <- check_convergence(summary, min_yr = 101, max_yr = 106)
   )
   expect_s3_class(ssb_check, "data.frame")
-  expect_true(nrow(ssb_check) == 4*3+4*6) # based on 2 scen. with 2 iter each, 2 tot em runs with 3 and 6 yrs.
[…]
