+++ man/run_SSMSE_scen.Rd
@@ -56,7 +56,7 @@ or \code{NULL}.}
 \item{EM_in_dir}{Relative or absolute path to the estimation model,}
 
 \item{run_EM_last_yr}{Should the MS be implemented to get future catch if the
-laste year is an assessment year? TRUE/FALSE option, so the same for all 
+last year is an assessment year? TRUE/FALSE option, so the same for all 
 scenarios and iterations. Defaults to FALSE.}
 
 \item{MS}{The management strategy to use. Current options
@@ -72,7 +72,10 @@ path to the file containing the function in `custom_MS_source`.}
 
 \item{custom_MS_source}{A file location with the source code for any custom MS
 models to be used in the simulation. SSMSE will source this file which should
-contain a function/s whose name/s match each custom MS models included in MS_vec.}
+contain a function/s whose name/s match each custom MS models included in MS_vec.
+To learn more about using custom management strategies, see the 
+[using a Custom Management Stratey/Procedure section](https://nmfs-fish-tools.github.io/SSMSE/manual/advanced-options-use-a-custom-management-strategyprocedure.html)
+in the [SSMSE User Manual](https://nmfs-fish-tools.github.io/SSMSE/manual).}
 
 \item{use_SS_boot}{Should a bootstrapped data set generated by SS be used?
 Defaults to TRUE.}
