### 2023-09-06 #4937 @k-doering-NOAA
@quang-huynh I realized the changes we made were in a branch, rather than in the main branch, from which the paper draft was being built. I just went and merged the changes into the main branch and rebuilt the paper - see the above [editorialbot comment](https://github.com/openjournals/joss-reviews/issues/4937#issuecomment-1708519417) that should include the changes to the draft. I hope this helps!
