### 2023-09-06 #4937 @quang-huynh
Thanks @k-doering-NOAA

The updates are suitable for me and I've checked off the outstanding items.

Some minor comments:

> > **Functionality documentation** Projected dynamics in the operating model can be tinkered with a list, with a template provided by `SSMSE::create_future_om_list`. However, there's no documentation in the help file on what information is needed and what options are possible. What are the required values for 'pars', 'scen', 'pattern', and 'input' entries in the list? I can figure it out with pattern recognition from the examples but that limits me to the case study.
>
> It’s a really good point - we added a link to the SSMSE user manual in the R help documentation ([nmfs-fish-tools/SSMSE@1452b0a](https://github.com/nmfs-fish-tools/SSMSE/commit/1452b0a)).

The additions to the manual for the om_list and sample_struct are quite helpful. The link in the R help isn't working as roxygen doesn't use markdown formatting, I believe the correct format is `\href{https://}{text}`.

Not required for the paper revision, but I'd also recommend using the term 'projection' as much as possible in the manual when referencing the simulations. "Future changes to the OM" sounds like it refers to an update in the functionality of the software in a later version.

**Paper comments:**

Line 32: I don't like using the word "realistic" here as all models are simplifications of the real world. Recommend "developing sufficient OMs.." or "developing sufficient OMs for evaluating management strategies is intensive" and leave it at that.

Line 76: "file directories" instead of "names and locations"?

Line 106: Assessment error is also an emergent property of the estimation algorithm, e.g., retrospective bias. The simulation itself, i.e., model fitting in the projection, allows one to replicate these dynamics as part of the management procedure evaluation.

Line 122: Markdown formatting issue.
