### 2023-09-08 1e615d3
Paper edits suggested by @quang-huynh
in https://github.com/openjournals/joss-reviews/issues/4937#issuecomment-1708899444
