### 2023-09-08 e80a9e1
fix: generalize startup message
