+++ R/zzz.R
@@ -1,3 +1,5 @@
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Welcome to SSMSE version 0.2.6 using Stock Synthesis 3.30.18.")
+  packageStartupMessage("Welcome to SSMSE version ",
+  packageVersion(pkgname),
+  " using Stock Synthesis 3.30.18.")
 }
