+++ man/SSMSE_summary_all.Rd
@@ -31,11 +31,11 @@ available)}
 }
 \value{
 A list of 3 data frames called scalar, ts, and
- dq (for derived quantities). These lists contain information for
- multiple model runs (estimation models and operating models) for 1
- iteration.Also writes 3 .csv files with the contents of this list of
- dataframes to dir and 3.csv files with scenario specific results in each of
- the scenario foldurs..
+dq (for derived quantities). These lists contain information for
+multiple model runs (estimation models and operating models) for 1
+iteration.Also writes 3 .csv files with the contents of this list of
+dataframes to dir and 3.csv files with scenario specific results in each of
+the scenario foldurs..
 }
 \description{
 Get results in a list for 1 iteration, using ss3sim::get_results_iter
