+++ man/run_SSMSE_iter.Rd
@@ -48,26 +48,26 @@ or \code{NULL}.}
 \item{EM_in_dir}{Relative or absolute path to the estimation model,}
 
 \item{run_EM_last_yr}{Should the MS be implemented to get future catch if the
-last year is an assessment year? TRUE/FALSE option, so the same for all 
+last year is an assessment year? TRUE/FALSE option, so the same for all
 scenarios and iterations. Defaults to FALSE.}
 
 \item{MS}{The management strategy to use. Current options
-are:  \code{"last_yr_catch"} which uses the previous year's catch; 
-\code{"no_catch"} which uses 0 catch; \code{"EM"} which uses an stock 
-synthesis model as the estimation method and the management strategy as 
+are:  \code{"last_yr_catch"} which uses the previous year's catch;
+\code{"no_catch"} which uses 0 catch; \code{"EM"} which uses an stock
+synthesis model as the estimation method and the management strategy as
 defined in the forecast file of the stock synthesis estimation method;
 \code{"Interim"} to modify catch based on survey predictions between
 assessments. Users can also specify their own management strategies as a
 function. For example, if the function is
-called "my_ms" then the user should specify MS = "my_ms" and specify the 
-path to the file containing the function in `custom_MS_source`.}
+called "my_ms" then the user should specify MS = "my_ms" and specify the
+path to the file containing the function in \code{custom_MS_source}.}
 
 \item{custom_MS_source}{A file location with the source code for any custom MS
 models to be used in the simulation. SSMSE will source this file which should
 contain a function/s whose name/s match each custom MS models included in MS_vec.
-To learn more about using custom management strategies, see the 
-[using a Custom Management Stratey/Procedure section](https://nmfs-fish-tools.github.io/SSMSE/manual/advanced-options-use-a-custom-management-strategyprocedure.html)
-in the [SSMSE User Manual](https://nmfs-fish-tools.github.io/SSMSE/manual).}
+To learn more about using custom management strategies, see the
+\href{https://nmfs-fish-tools.github.io/SSMSE/manual/custom-MS.html}{using a Custom Management Stratey/Procedure section}
+in the \href{https://nmfs-fish-tools.github.io/SSMSE/manual}{SSMSE User Manual}.}
 
 \item{use_SS_boot}{Should a bootstrapped data set generated by SS be used?
 Defaults to TRUE.}
@@ -117,7 +117,7 @@ sampled for the historical period for the data generated from the OM. If
 this is left as NULL, then the same sampling scheme will be used as in the
 OM's data file. If it is not NULL, then each year.}
 
-\item{sample_catch}{Should catch be sampled or fixed at the OM values? 
+\item{sample_catch}{Should catch be sampled or fixed at the OM values?
 Defaults to FALSE.}
 
 \item{interim_struct}{A optional list of parameters to control an interim
@@ -128,14 +128,14 @@ number of years between full assessments during with an interim assessment
 will happen every year, and Index_weights is a vector of length n indexes
 that weights all indexes for multi index inference.
 interim_struct<-list(Beta = 1,
-                     MA_years = 3,
-                     assess_freq = 5,
-                     Index_weights = rep(1, max(ref_index[, 3])))}
+MA_years = 3,
+assess_freq = 5,
+Index_weights = rep(1, max(ref_index\link{, 3})))}
 
 \item{n_F_search_loops}{Number of times to try to find an F that achieves the
 catches input in the OM. Defaults to 20.}
 
-\item{tolerance_F_search}{How far apart the input catch and achieved 
+\item{tolerance_F_search}{How far apart the input catch and achieved
 catch can be in tried to find an F that achieves the catch input in the
 OM. Defaults to 0.001.}
 

[…]
