### 2023-09-08 74407b7
manual and r oxygen fixes use markdown syntax in roxygen, fix: manual crosslinks, use projection instead of future changes
