+++ man/set_MSE_seeds.Rd
@@ -8,11 +8,11 @@ set_MSE_seeds(seed = NULL, iter_vec)
 }
 \arguments{
 \item{seed}{Input a fixed seed to replicate previous simulation runs. Seed
-can be a single value for a global seed, n_scenarios+1 length vector for 
-scenario specific and a global seed, n_iterations+n_scenarios+1 length 
+can be a single value for a global seed, n_scenarios+1 length vector for
+scenario specific and a global seed, n_iterations+n_scenarios+1 length
 vector for iteration scenario and global seeds. Can also be a list object
-with a single value under \code{seed[["global"]]}, a vector under 
-\code{seed[["scenario"]]}, and a multiple vectors for iteration specific seeds 
+with a single value under \code{seed[["global"]]}, a vector under
+\code{seed[["scenario"]]}, and a multiple vectors for iteration specific seeds
 under \code{seed[["iter"]][[1:n_scenarios]]}.}
 
 \item{iter_vec}{The number of iterations per scenario. A vector of integers
