+++ man/SSMSE_summary_scen.Rd
@@ -12,10 +12,10 @@ Defaults to the working directory.}
 }
 \value{
 A list of 3 data frames called scalar, ts, and
- dq (for derived quantities). These lists contain information for
- multiple model runs (estimation models and operating models) for 1
- iteration.Also writes 3 .csv files with the contents of this list of
- dataframes to dir.
+dq (for derived quantities). These lists contain information for
+multiple model runs (estimation models and operating models) for 1
+iteration.Also writes 3 .csv files with the contents of this list of
+dataframes to dir.
 }
 \description{
 Get results in a list for 1 iteration, using ss3sim::get_results_iter
