+++ man/SSMSE_summary_iter.Rd
@@ -11,9 +11,9 @@ SSMSE_summary_iter(dir)
 }
 \value{
 A list of 3 data frames called scalar, timeseries, and
- derived (for derived quantities). These lists contain information for
- multiple model runs (estimation models and operating models) for 1
- iteration.
+derived (for derived quantities). These lists contain information for
+multiple model runs (estimation models and operating models) for 1
+iteration.
 }
 \description{
 Get results in a list for 1 iteration, using ss3sim::get_results_iter
