### 2023-09-08 38de94c
style and docs: run devtools::document() and styler::style_pkg()
