+++ R/zzz.R
@@ -1,5 +1,7 @@
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Welcome to SSMSE version ",
-  packageVersion(pkgname),
-  " using Stock Synthesis 3.30.18.")
+  packageStartupMessage(
+    "Welcome to SSMSE version ",
+    packageVersion(pkgname),
+    " using Stock Synthesis 3.30.18."
+  )
 }
