### 2023-09-08 1000ed3
style and docs: run devtools::document() and styler::style_pkg()
