+++ DESCRIPTION
@@ -39,5 +39,5 @@ Remotes:
     ss3sim/ss3sim@main
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
 Roxygen: list(markdown = TRUE)
+RoxygenNote: 7.2.3
