### 2023-09-08 #4937 @k-doering-NOAA
thanks @quang-huynh , I added your changes to the paper, which should show up in the newly generated version.

Thanks for the manual suggestions, too, I turned on markdown formatting for Roxygen so that it renders properly.
