### 2023-09-13 nmfs-fish-tools/SSMSE#152 @k-doering-NOAA
I was also recently made aware of [parallelly](https://parallelly.futureverse.org/#availablecores-vs-paralleldetectcores) which offers a enhanced option of `parallel::detectCores()` (`parallelly::availableCores()`)
