### 2023-09-27 6cdf1a5
copy edits
additional edits to equation formatting
