### 2023-09-30 #4937 @k-doering-NOAA
@sbenthall the tag is `v0.2.7` (https://github.com/nmfs-fish-tools/SSMSE/tree/v0.2.7)

We proofread the draft and made small fixes. I noticed Sean Anderson was listed as a reviewer, even though he ultimately did not review - I don't know how to fix this.

I'll work on Zonodo now.
