### 2023-10-14 #4937 @arfon
@sbenthall – I wonder if it's [this line](https://github.com/nmfs-fish-tools/SSMSE/blob/main/JOSS/paper.bib#L35), where the DOI is double-wrapped `{{ doi }}`. Perhaps update it to be the same as others `{ doi }`
