### 2023-10-16 #4937 @sbenthall
@k-doering-NOAA Hello. It looks like a formatting error in your paper.bib file is blocking the robot from accepting you paper! Can you please correct this?

After making that change, please create a new release and archive DOI and tell me, so that everything is synced up. :)
