### 2023-10-16 1a94b9a
Fix: bib file issues
