+++ JOSS/paper.bib
@@ -32,7 +32,7 @@
 
   year = {2006},
 
-  doi = {{10.1577/1548-8446(2006)31[590:TCFIE]2.0.CO;2}},
+  doi = {10.1577/1548-8446(2006)31[590:TCFIE]2.0.CO;2},
 
   publisher={Wiley Online Library}
 
@@ -117,11 +117,11 @@ eprint = {https://www.science.org/doi/pdf/10.1126/science.1149016}
 
  title={Report for the first north Pacific albacore management strategy evaluation},
 
- author={{International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean}},
+ author={International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean},
 
  year={2019},
 
- institution={{International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean}},
+ institution={International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean},
 
  type ={annex 12},
 
@@ -171,7 +171,7 @@ keywords = {Weakfish, Tagging, Fishing mortality, Natural mortality, Catch-curve
 
 @article{maceetal2021,
 
-title = {{Spatial population dynamics of eastern oyster in the Chesapeake Bay, Maryland}},
+title = {Spatial population dynamics of eastern oyster in the Chesapeake Bay, Maryland},
 
 journal = {Fisheries Research},
 
@@ -193,7 +193,7 @@ author = {Marvin M. Mace and Kathryn L. Doering and Michael J. Wilberg and Amy L
 
 author = {Marty, Gary D and Quinn, Terrance J and Carpenter, Greg and Meyers, Theodore R and Willits, Neil H},
 
-title = {{Role of disease in abundance of a Pacific herring (\textit{Clupea pallasi}) population}},
+title = {Role of disease in abundance of a Pacific herring (\textit{Clupea pallasi}) population},
 
 journal = {Canadian Journal of Fisheries and Aquatic Sciences},
 
@@ -267,7 +267,7 @@ URL = { https://doi.org/10.1139/f03-109}
 
 @article{methotandwetzel2013,
 
-  title={{Stock Synthesis: a biological and statistical framework for fish stock assessment and fishery management}},
+  title={Stock Synthesis: a biological and statistical framework for fish stock assessment and fishery management},
 
   author={Methot, Richard D and Wetzel, Chantel R},
 
@@ -315,7 +315,7 @@ URL = { https://doi.org/10.1139/f03-109}
 
 @misc{stocksmart,
 
-  author = {{NOAA Fisheries}},
+  author = {NOAA Fisheries},
 
   title = {Stock SMART data records},
 
@@ -501,7 +501,7 @@ keywords = {Northern cod, Prey limitations, Body condition, Starvation, Mortalit
 
 @article{steidinger2009,
 
-title = {{Historical perspective on \textit{Karenia brevis} red tide research in the Gulf of Mexico}},
+title = {Historical perspective on \textit{Karenia brevis} red tide research in the Gulf of Mexico},
 
 journal = {Harmful Algae},
 
