### 2023-10-23 #4937 @kthyng
@k-doering-NOAA We usually ask that authors change the metadata in the zenodo archive to match the JOSS title and author list — would you mind doing this?
