### 2023-10-23 #4937 @kthyng
@k-doering-NOAA
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
