### 2023-10-24 #4937 @k-doering-NOAA
@kthyng I just changed the Zenodo metadata title to match the JOSS title. Author list already matches.

For the references, do I need to edit the .bib file and resubmit to Zenodo again?
