### 2023-10-24 #4937 @kthyng
Yes edit the .bib and let's regenerate the pdf to check everything. No need to resubmit to Zenodo since it is just the paper text changing.
