### 2023-10-24 5537bc5
fix capitals, authors use {} to lock case or fix author name
