+++ JOSS/paper.bib
@@ -461,7 +461,7 @@ keywords = {Northern cod, Prey limitations, Body condition, Starvation, Mortalit
 
 @article{sharmaetal2020,
 
-  title={Operating model design in tuna {R}egional {F}ishery {M}anagement Organizations: Current practice, issues and implications},
+  title={Operating model design in tuna {R}egional {F}ishery {M}anagement {O}rganizations: Current practice, issues and implications},
 
   author={Sharma, Rishi and Levontin, Polina and Kitakado, Toshihide and Kell, Laurence and Mosqueira, Iago and Kimoto, Ai and Scott, Rob and Minte-Vera, Carolina and De Bruyn, Paul and Ye, Yimin and others},
 
