### 2023-10-24 856fba1
Fix bibliograph for JOSS paper (#159)
* fix capitals, authors
use {} to lock case or fix author name

* fix: one additional capital
