+++ JOSS/paper.bib
@@ -1,7 +1,7 @@
 
  @Article{andersonetal2014,
 
-  title = {ss3sim: An R package for fisheries stock assessment simulation with Stock Synthesis},
+  title = {ss3sim: An {R} package for fisheries stock assessment simulation with {S}tock {S}ynthesis},
   author = {Sean C. Anderson and Cole C. Monnahan and Kelli F. Johnson and Kotaro Ono and Juan L. Valero},
   year = {2014},
 
@@ -19,7 +19,7 @@
 
 @article{brownetal2006,
 
-  title = {The Center for Independent Experts: The national external peer review program of NOAA's National Marine Fisheries Service},
+  title = {The {C}enter for {I}ndependent {E}xperts: The national external peer review program of {NOAA}'s {N}ational {M}arine {F}isheries {S}ervice},
   author = {Brown, Stephen K and Shivlani, Manoj and Die, David and Sampson, David B and  Ting, Tina A},
 
   journal = {Fisheries},
@@ -42,7 +42,7 @@
 
 author = {F. Chan  and J. A. Barth  and J. Lubchenco  and A. Kirincich  and H. Weeks  and W. T. Peterson  and B. A. Menge },
 
-title = {Emergence of Anoxia in the California Current Large Marine Ecosystem},
+title = {Emergence of Anoxia in the {C}alifornia {C}urrent {L}arge {M}arine {E}cosystem},
 
 journal = {Science},
 
@@ -104,7 +104,7 @@ eprint = {https://www.science.org/doi/pdf/10.1126/science.1149016}
 
 @Manual{openMSEcite,
 
-    title = {openMSE: Easily Install and Load the 'openMSE' Packages},
+    title = {{openMSE}: Easily Install and Load the {'openMSE'} Packages},
 
     author = {Adrian Hordyk and Quang Huynh and Tom Carruthers},
 
@@ -115,15 +115,13 @@ eprint = {https://www.science.org/doi/pdf/10.1126/science.1149016}
 
 @techreport{isc2019,
 
- title={Report for the first north Pacific albacore management strategy evaluation},
+ title={Report for the first north {P}acific albacore {M}anagement {S}trategy {E}valuation},
 
- author={International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean},
+ author={{International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean}},
 
  year={2019},
 
- institution={International Scientific Committee for Tuna and Tuna-Like Species in the North Pacific Ocean},
-
- type ={annex 12},
+ type ={Annex 12},
 
  URL={http://isc.fra.go.jp/pdf/ISC19/ISC19_ANNEX12_Report_First_North_Pacific_Albacore_MSE.pdf},
 
@@ -155,7 +153,7 @@ keywords = {Weakfish, Tagging, Fishing mortality, Natural mortality, Catch-curve
 
 @techreport{lynchetal2018,
 
-  title={Implementing a Next Generation Stock Assessment Enterprise: An Update to the NOAA Fisheries Stock Assessment Improvement Plan},
+  title={Implementing a Next Generation Stock Assessment Enterprise: An Update to the {NOAA} {F}isheries {S}tock {A}ssessment {I}mprovement {P}lan},
 
   author={Lynch, Patrick D and Methot, Richard D and Link, Jason S},
 
@@ -171,7 +169,7 @@ keywords = {Weakfish, Tagging, Fishing mortality, Natural mortality, Catch-curve
 
 @article{maceetal2021,
 
-title = {Spatial population dynamics of eastern oyster in the Chesapeake Bay, Maryland},
+title = {Spatial population dynamics of eastern oyster in the {C}hesapeake {B}ay, {M}aryland},
 
[…]
