+++ JOSS/paper.bib
@@ -191,7 +191,7 @@ author = {Marvin M. Mace and Kathryn L. Doering and Michael J. Wilberg and Amy L
 
 author = {Marty, Gary D and Quinn, Terrance J and Carpenter, Greg and Meyers, Theodore R and Willits, Neil H},
 
-title = {Role of disease in abundance of a {P}acific herring (\textit{Clupea pallasi}) population},
+title = {Role of disease in abundance of a {P}acific herring (\textit{{C}lupea pallasi}) population},
 
 journal = {Canadian Journal of Fisheries and Aquatic Sciences},
 
