### 2023-10-24 #4937 @k-doering-NOAA
@kthyng, thank you, I made changes to the bib and pushed them to the main branch (https://github.com/nmfs-fish-tools/SSMSE). I used the JOSS paper draft github action to generate an informal proof and I think the capitalization issue is fixed. Thanks for the tip about `{}`!
