### 2023-10-26 89b8655
update ss3 repo names; @k-doering-NOAA the model files will need updating after new release
