### 2023-10-26 0db1929
update ss3 repo names
