### 2023-10-26 586109d
modify remotes in description because assertive repos are no longer on CRAN.
