+++ DESCRIPTION
@@ -36,7 +36,10 @@ Suggests:
     testthat
 Remotes:
     r4ss/r4ss@main,
-    ss3sim/ss3sim@main
+    ss3sim/ss3sim@main,
+    cran/assertive.base@master,
+    cran/assertive.properties@master,
+    cran/assertive.types@master
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
