### 2023-10-26 a0e9c53
fixes caught by R cmd check
