### 2023-10-26 5a64210
fixes caught by R cmd check
