+++ R/zzz.R
@@ -1,7 +1,7 @@
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(
     "Welcome to SSMSE version ",
-    packageVersion(pkgname),
+    utils::packageVersion(pkgname),
     " using Stock Synthesis 3.30.18."
   )
 }
