+++ man/run_SSMSE.Rd
@@ -122,7 +122,7 @@ MA_years= the number of years to average index observations of when calculating 
 assess_freq=the number of years between full assessments during with an interim assessment will happen
 every year, and Index_weights is a vector of length n indexes that weights all indexes for multi index
 inference.
-interim_struct_list<-list(Beta=1,MA_years=3,assess_freq=5,Index_weights=rep(1,max(ref_index\link{,3})))}
+\code{interim_struct_list<-list(Beta=1,MA_years=3,assess_freq=5,Index_weights=rep(1,max(ref_index[,3])))}}
 
 \item{verbose}{Want verbose output? Defaults to FALSE.}
 
