### 2023-10-30 e9d1738
style and docs: run devtools::document() and styler::style_pkg()
