+++ JOSS/paper.bib
@@ -499,7 +499,7 @@ keywords = {Northern cod, Prey limitations, Body condition, Starvation, Mortalit
 
 @article{steidinger2009,
 
-title = {Historical perspective on \textit{Karenia brevis} red tide research in the {G}ulf of {M}exico},
+title = {Historical perspective on \textit{{K}arenia brevis} red tide research in the {G}ulf of {M}exico},
 
 journal = {Harmful Algae},
 
